AI/Vaccination_API/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
google-generativeai==0.4.1
google-cloud-translate==3.12.1
